Support remotes not ending in .git
The .git suffix is entirely unnecessary, and I am not in the habit of
having it; I see no reason why this project should require it.

The critical part in the regexParseProjectName change is anchoring the
match to the end of the string (`$`) and making the project name
matching non-greedy (adding `?` after `.+`) so that an optional `.git`
would consume that part, rather than the project name including `.git`.
But I also substantially improved the regular expression while I was at
it to make it much more robust.